feat: 수량카드-구매자/판매자 구현
